cleanup: AsyncStreamingReadWriteRpc needs to be public
